Revert "More commands (#74)"
This reverts commit c3d76d4567add006e68d925b41b308236c5b273b.
