+++ package.json
@@ -7,7 +7,7 @@
 		"start": "npm run prod",
 		"build": "npm-run-all clean transpile",
 		"server": "node ./dist-server/index.js",
-		"dev": "nodemon --exec babel-node src",
+		"dev": "npm-run-all build server",
 		"prod": "npm-run-all build server",
 		"transpile": "babel ./src --out-dir dist-server --copy-files",
 		"clean": "rimraf dist-server",
@@ -51,7 +51,6 @@
 		"cors": "^2.8.5",
 		"crypto": "^1.0.1",
 		"date-fns": "^2.14.0",
-		"dblapi.js": "^2.4.1",
 		"deep-equal": "^2.0.4",
 		"discord-anti-spam": "^2.4.1",
 		"discord-oauth2": "^2.2.0",
@@ -90,7 +89,6 @@
 		"uuidv1": "^1.6.14"
 	},
 	"devDependencies": {
-		"@babel/node": "^7.12.6",
 		"@babel/cli": "^7.10.5",
 		"@babel/core": "^7.11.0",
 		"@babel/plugin-transform-runtime": "^7.11.0",
