+++ src/Discord/Commands/CustomCommands/index.js
@@ -44,7 +44,7 @@ module.exports = async ({ command, args, message, client }) => {
 				// check if the user can use this command based on their roles
 				const roles = message.member.roles;
                 const roleIds = roles.cache.array().map(role => role.id);
-                // console.log(value.permittedRoles, roleIds)
+                console.log(value.permittedRoles, roleIds)
 				if (value.permittedRoles) {
 					if (!ArrayAny(value.permittedRoles, roleIds)) {
                         console.log("allowed")
