+++ src/Discord/Commands/admin/restart.js
@@ -8,10 +8,9 @@ module.exports = {
     description: "Restarts the Client.",
 	adminOnly: true,
 	execute: async (message, args, client) => {
-		const token = client.token
 		const msg = await message.channel.send("Restarting...");
 		await client.destroy();
-		await client.login(token);
-		await msg.edit("Restarted");
+		await client.login(process.env.BOT_TOKEN);
+		await message.channel.send("Restarted");
 	},
 };
