+++ src/Discord/Commands/info/help.js
@@ -1,6 +1,6 @@
 const { isAdmin, hasPermission, ArrayAny, getRoleIds } = require("../../../utils/functions");
 const { MessageEmbed } = require("discord.js");
-import { getDiscordSettings, convertDiscordRoleColor } from "../../../utils/functions";
+import { getDiscordSettings } from "../../../utils/functions";
 
 // the admin app has already been initialized in routes/index.js
 const admin = require("firebase-admin");
@@ -18,7 +18,7 @@ const getCommands = async (message, client, plugins) => {
 		} else if (command.adminOnly && isAdmin(message.author)) {
 			availableCommands.push(commandObj);
 			continue;
-		} else if (command.permissions?.length && (await hasPermission(message.member, command.permissions || []))) {
+		} else if (command.permissions?.length && await hasPermission(message.member, command.permissions || [])) {
 			availableCommands.push(commandObj);
 			continue;
 		}
@@ -55,57 +55,6 @@ const filterCustomCommands = (commands, { member, channel }) => {
 	}
 	return availableCommands;
 };
-
-const addTips = async (embed, msg, client) => {
-	embed.addField(
-		"Tip",
-		"Type `help <command name>` for help on a specific commands and `help commands <command name>` to get help on a specific custom command "
-	);
-	// if (await hasPermission(message.member, ["MANAGE_SERVER", "ADMINISTRATOR"])) {
-	// 	helpEmbed.addField(
-	// 		"Moderator Tip",
-	// 		"Type `help module` to get informations about the available module or `help module <module name>` for help on a specific module"
-	// 	);
-	// }
-	// if (isAdmin(message.author)) {
-	// 	helpEmbed.addField("DisStreamChat Admin Tip", "Type `help admin` for links to DisStreamChat admin tools");
-	// }
-	embed.addField("Support Server", "If you have any questions or bug reports come tell us at http://discord.disstreamchat.com");
-	embed.addField("Custom Commands", "To get more help on custom commands use `help commands`");
-	return embed;
-};
-
-const maxCommands = 2;
-const generateHelpEmbed = async ({ message, client, commands, custom, page = 1 }) => {
-	const pages = custom ? Math.ceil(commands.length / (maxCommands * 4)) : Math.ceil(Object.keys(commands || {}).length / maxCommands);
-	const index = (page - 1) * maxCommands;
-	const helpEmbed = new MessageEmbed()
-		.setTitle("DisStreambot Help")
-		.setDescription(`Here are all the available commands.`)
-		.addField("Prefix", client.prefix || "!")
-		// .setThumbnail(client.user.displayAvatarURL())
-		.setAuthor("DisStreamBot Commands", client.user.displayAvatarURL())
-		.setTimestamp(message.createdAt)
-		.setFooter(`Page ${page}/${pages}`)
-		.setColor(convertDiscordRoleColor(message.guild.me.displayHexColor))
-		.addField(
-			"Available Commands",
-			custom
-				? commands
-						.slice(index, index + maxCommands*4)
-						.map(command => (command.displayName.includes("<:") ? command.displayName : `\`${command.displayName}\``))
-						.join(", ")
-				: Object.entries(commands)
-						.slice(index, index + maxCommands)
-						.map(([key, value]) => {
-							return `**${key}**\n${value.map(command => `\`${command.displayName}\``).join(", ")}\n`;
-						})
-		)
-		.addField("---------------------------------------------------", "--------------------------------------------------");
-
-	return { maxPages: pages, embed: await addTips(helpEmbed) };
-};
-
 module.exports = {
 	name: "help",
 	id: "help",
@@ -120,55 +69,36 @@ module.exports = {
 		const guildData = guildRef.data();
 		let customCommands = filterCustomCommands(guildData, message);
 		const allCommands = [...availableCommands, ...customCommands];
-		const commandCategories = availableCommands.reduce((categories, current) => {
-			if (categories[current.category]) categories[current.category].push(current);
-			else categories[current.category] = [current];
-			return categories;
-		}, {});
-		const allCommandCategories = { ...commandCategories, custom: customCommands };
 		if (args.length === 0) {
-			const { embed: helpEmbed, maxPages } = await generateHelpEmbed({ message, client, commands: allCommandCategories, page: 1 });
-			const helpMsg = await message.channel.send(helpEmbed);
-			const pageCollector = helpMsg.createReactionCollector(
-				(reaction, user) => ["⬅️", "➡️", "❌"].includes(reaction.emoji.name) && !user.bot
+			const helpEmbed = new MessageEmbed()
+				.setTitle("DisStreambot Help")
+				.setDescription(`Here are all the available commands.`)
+				.addField("Prefix", client.prefix || "!")
+				.setThumbnail(client.user.displayAvatarURL())
+				.setAuthor("DisStreamBot Commands", client.user.displayAvatarURL())
+				.addField("Available Commands", allCommands.map(command => `\`${command.displayName}\``).join(", "))
+				.addField(
+					"Tip",
+					"Type `help <command name>` for help on a specific commands and `help commands <command name>` to get help on a specific custom command "
+				)
+				.setTimestamp(message.createdAt)
+				.setColor("#206727");
+			if (await hasPermission(message.member, ["MANAGE_SERVER", "ADMINISTRATOR"])) {
+				helpEmbed.addField(
+					"Moderator Tip",
+					"Type `help module` to get informations about the available module or `help module <module name>` for help on a specific module"
+				);
+			}
+			if (isAdmin(message.author)) {
+				helpEmbed.addField("DisStreamChat Admin Tip", "Type `help admin` for links to DisStreamChat admin tools");
+			}
+			helpEmbed.addField(
+				"Support Server",
+				"If you have any questions or bug reports come tell us at http://discord.disstreamchat.com"
 			);
-			let currentPage = 1;
-			await helpMsg.react("⬅️");
-			await helpMsg.react("➡️");
-			await helpMsg.react("❌");
-			pageCollector.on("end", () => {
-				helpMsg.reactions.removeAll();
-			});
-			pageCollector.on("collect", async reaction => {
-				let pageChanged = false;
-				switch (reaction.emoji.name) {
-					case "⬅️":
-						if (currentPage > 1) {
-							currentPage -= 1;
-							pageChanged = true;
-						}
-						break;
-					case "➡️":
-						if (currentPage < maxPages) {
-							currentPage += 1;
-							pageChanged = true;
-						}
-						break;
-					case "❌":
-						pageCollector.stop();
-						break;
-					default:
-						return;
-				}
-				const senders = reaction.users.cache.array().filter(user => user.id !== client.user.id);
-				for (const sender of senders) {
-					reaction.users.remove(sender);
-				}
-				if (pageChanged) {
-					helpMsg.edit((await generateHelpEmbed({ message, client, commands: allCommandCategories, page: currentPage })).embed);
-				}
-			});
-		} else if (!["module", "admin", "commands"].includes(args[0])) {
+			helpEmbed.addField("Custom Commands", "To get more help on custom commands use `help commands`");
+			await message.channel.send(helpEmbed);
+		} else if (args[0] !== "module" && args[0] !== "admin" && args[0] !== "commands") {
 			const selectedCommand = allCommands.find(command => command.displayName?.toLowerCase() === args[0]?.toLowerCase());
 			const commandHelpEmbed = getHelpText({ message, client, selectedCommand });
 			if (!commandHelpEmbed) {
@@ -188,63 +118,27 @@ module.exports = {
 						}
 					} else {
 						const availableCustomCommands = await getCommands(message, { commands: guildData });
-						const { embed: customHelpEmbed, maxPages } = await generateHelpEmbed({
-							message,
-							client,
-							commands: availableCustomCommands,
-							custom: true,
-							page: 1,
-						});
-						const helpMsg = await message.channel.send(customHelpEmbed);
-						const pageCollector = helpMsg.createReactionCollector(
-							(reaction, user) => ["⬅️", "➡️", "❌"].includes(reaction.emoji.name) && !user.bot
-						);
-						let currentPage = 1;
-						await helpMsg.react("⬅️");
-						await helpMsg.react("➡️");
-						await helpMsg.react("❌");
-						pageCollector.on("end", () => {
-							helpMsg.reactions.removeAll();
-						});
-						pageCollector.on("collect", async reaction => {
-							let pageChanged = false;
-							switch (reaction.emoji.name) {
-								case "⬅️":
-									if (currentPage > 1) {
-										currentPage -= 1;
-										pageChanged = true;
-									}
-									break;
-								case "➡️":
-									if (currentPage < maxPages) {
-										currentPage += 1;
-										pageChanged = true;
-									}
-									break;
-								case "❌":
-									pageCollector.stop();
-									break;
-								default:
-									return;
-							}
-							const senders = reaction.users.cache.array().filter(user => user.id !== client.user.id);
-							for (const sender of senders) {
-								reaction.users.remove(sender);
-							}
-							if (pageChanged) {
-								helpMsg.edit(
-									(
-										await generateHelpEmbed({
-											message,
-											client,
-											commands: availableCustomCommands,
-											custom: true,
-											page: currentPage,
-										})
-									).embed
-								);
-							}
-						});
+						const customHelpEmbed = new MessageEmbed()
+							.setTitle("DisStreambot Help")
+							.setDescription("Here are all the available custom commands")
+							.setThumbnail(client.user.displayAvatarURL())
+							.setAuthor("DisStreamBot Commands", client.user.displayAvatarURL())
+							.addField(
+								"Available Commands",
+								availableCustomCommands
+									.map(command =>
+										command.displayName.includes("<:") ? command.displayName : `\`${command.displayName}\``
+									)
+									.join(", ")
+							)
+							.addField("Tip", "Type `help commands <command name>` to get help on a specific command ")
+							.setTimestamp(message.createdAt)
+							.setColor("#206727")
+							.addField(
+								"Support Server",
+								"If you have any questions or bug reports come tell us at http://discord.disstreamchat.com"
+							);
+						await message.channel.send(customHelpEmbed);
 					}
 					break;
 			}
